+++ src/main/java/codesquad/fineants/domain/gainhistory/service/PortfolioGainHistoryService.java
@@ -4,7 +4,6 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -25,13 +24,6 @@ public class PortfolioGainHistoryService {
 	private final PortfolioGainHistoryRepository repository;
 	private final PortfolioRepository portfolioRepository;
 	private final CurrentPriceRedisRepository currentPriceRedisRepository;
-
-	@Transactional
-	@Scheduled(cron = "0 0 16 * * ?") // 매일 16시에 실행
-	public void scheduledPortfolioGainHistory() {
-		PortfolioGainHistoryCreateResponse response = addPortfolioGainHistory();
-		log.info("포트폴리오 수익 내역 기록 결과, size = {}", response.getIds().size());
-	}
 
 	@Transactional
 	public PortfolioGainHistoryCreateResponse addPortfolioGainHistory() {
