+++ src/main/java/codesquad/fineants/domain/kis/scheduler/KisClosingPriceScheduler.java
@@ -2,9 +2,8 @@ package codesquad.fineants.domain.kis.scheduler;
 
 import java.time.LocalDate;
 
-import org.springframework.context.annotation.Profile;
 import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import codesquad.fineants.domain.kis.repository.HolidayRepository;
@@ -12,23 +11,20 @@ import codesquad.fineants.domain.kis.service.KisService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
-@Profile(value = "production")
-@Slf4j
+@Component
 @RequiredArgsConstructor
-@Service
-public class KisProductionScheduledService {
-
-	private final HolidayRepository holidayRepository;
+@Slf4j
+public class KisClosingPriceScheduler {
 	private final KisService kisService;
+	private final HolidayRepository holidayRepository;
 
-	@Scheduled(cron = "0/5 * 9-15 ? * MON,TUE,WED,THU,FRI")
-	@Transactional
-	public void refreshCurrentPrice() {
-		// 휴장일인 경우 실행하지 않음
+	// start pm 15:30
+	@Scheduled(cron = "${cron.expression.closing-price:* 30 15 * * *}")
+	@Transactional(readOnly = true)
+	public void scheduledRefreshAllClosingPrice() {
 		if (holidayRepository.isHoliday(LocalDate.now())) {
 			return;
 		}
-		kisService.refreshAllStockCurrentPrice();
+		kisService.refreshAllClosingPrice();
 	}
-
 }
