+++ src/main/resources/application-production.yml
@@ -25,3 +25,4 @@ server:
 cron:
   expression:
     reload-stocks: '0 0 8 * * ?'
+    closing-price: '* * 16 * * *'
